+++ samples/BeatPulseTrackers/Program.cs
@@ -12,6 +12,7 @@ namespace BeatPulseTrackers
 
         public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
            WebHost.CreateDefaultBuilder(args)
+               .UseApplicationInsights()
                .UseBeatPulse(setup =>
                {
                    setup.ConfigureDetailedOutput()