+++ src/BeatPulse.Oracle/OracleLiveness.cs
@@ -1,0 +1,40 @@
+﻿using BeatPulse.Core;
+using Microsoft.AspNetCore.Http;
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Oracle.ManagedDataAccess.Client;
+namespace BeatPulse.Oracle
+{
+    public class OracleLiveness : IBeatPulseLiveness
+    {
+        private readonly string _connectionString;
+
+        public OracleLiveness(string connectionString)
+        {
+            _connectionString = connectionString ?? throw new ArgumentNullException(nameof(connectionString));
+        }
+        public async Task<(string, bool)> IsHealthy(HttpContext context, LivenessExecutionContext livenessContext, CancellationToken cancellationToken = default)
+        {
+            try
+            {
+                using (var connection = new OracleConnection(_connectionString))
+                {
+                    await connection.OpenAsync();
+                    var command = connection.CreateCommand();
+                    command.CommandText = "SELECT * FROM V$VERSION";
+                    await command.ExecuteScalarAsync();
+
+                    return (BeatPulseKeys.BEATPULSE_HEALTHCHECK_DEFAULT_OK_MESSAGE, true);
+                }
+            }
+            catch (Exception ex)
+            {
+                var message = !livenessContext.IsDevelopment ? string.Format(BeatPulseKeys.BEATPULSE_HEALTHCHECK_DEFAULT_ERROR_MESSAGE, livenessContext.Name)
+                        : $"Exception {ex.GetType().Name} with message ('{ex.Message}')";
+
+                return (message, false);
+            }
+        }
+    }
+}