+++ src/BeatPulse.StatusPageTracker/HttpClientExtensions.cs
@@ -1,0 +1,22 @@
+﻿using System.Net.Http;
+using System.Threading.Tasks;
+
+namespace BeatPulse.StatusPageTracker
+{
+    static class HttpClientExtensions
+    {
+        public static async Task<HttpResponseMessage> PatchAsync(this HttpClient client, string requestUri, HttpContent content)
+        {
+            var method = new HttpMethod("PATCH");
+
+            var request = new HttpRequestMessage(method, requestUri)
+            {
+                Content = content
+            };
+
+            var response = await client.SendAsync(request);
+
+            return response;
+        }
+    }
+}